moved substratee-node storage calls to crate substatee-node-calls crate, worker checks the amount of registered workers without acting upon the result
