+++ substratee-node-calls/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "substratee-node-calls"
+version = "0.1.0"
+authors = ["clangenbacher <christian.langenbacher@scs.ch>"]
+edition = "2018"
+
+# make this an independant workspace
+[workspace]
+
+[dependencies]
+log = "0.4"
+parity-codec = { version = "3.0", default-features = false }
+primitive-types = { version = "0.2", default-features = false, features = ["codec"] }
+
+[dependencies.substrate-api-client]
+git = "https://github.com/scs/substrate-api-client"
+default-features = false
+
+[dependencies.my_node_runtime]
+git = "https://github.com/scs/substraTEE-node"
+package = "substratee-node-runtime"
+features = ["no_global_allocator"]
