+++ worker/Cargo.toml
@@ -42,6 +42,9 @@ sgx_crypto_helper = { git = "https://github.com/baidu/rust-sgx-sdk", default-fea
 [dependencies.substrate-api-client]
 git = "https://github.com/scs/substrate-api-client"
 
+[dependencies.substratee-node-calls]
+path = "../substratee-node-calls"
+
 [dependencies.my_node_runtime]
 git = "https://github.com/scs/substraTEE-node"
 package = "substratee-node-runtime"
